// src/model/tree.ts
export type HealthStatus =
  | "Healthy"
  | "Degraded"
  | "Progressing"
  | "Unknown"
  | "Suspended"
  | "Missing";

export interface Health {
  status: HealthStatus;
}

export interface NodeBase {
  group: string;
  kind: string;
  namespace: string;
  name: string;
  uid: string;
}

export interface Node extends NodeBase {
  version: string;
  parentRefs?: Node[];
  resourceVersion: string;
  health: Health;
  createdAt: string;
}

export interface Tree {
  nodes: Node[];
}
// ...
export interface SyncOperation {
  revision: string;
  prune: boolean;
  dryRun: boolean;
  resources?: SyncOperationResource[];
}

export interface RetryBackoff {
  duration: string;
  maxDuration: string;
  factor: number;
}

export interface RetryStrategy {
  limit: number;
  backoff: RetryBackoff;
}

export interface RollbackOperation {
  id: number;
  prune: boolean;
  dryRun: boolean;
}

export interface OperationInitiator {
  username: string;
  automated: boolean;
}

export interface Operation {
  sync: SyncOperation;
  initiatedBy: OperationInitiator;
}
// ...
export interface OperationState {
  operation: Operation;
  phase: OperationPhase;
  message: string;
  syncResult: SyncOperationResult;
  startedAt: models.Time;
  finishedAt: models.Time;
}
// ...
export interface Application {
  apiVersion?: string;
  kind?: string;
  metadata: models.ObjectMeta;
  spec: ApplicationSpec;
  status: ApplicationStatus;
  operation?: Operation;
  isAppOfAppsPattern?: boolean;
}

export type WatchType = "ADDED" | "MODIFIED" | "DELETED" | "ERROR";

export interface ApplicationWatchEvent {
  type: WatchType;
  application: Application;
}

export interface ComponentParameter {
  component: string;
  name: string;
  value: string;
}

export interface ApplicationDestination {
  /**
   * Server address of the destination cluster
   */
  server: string;
  /**
   * Namespace of the destination cluster
   */
  namespace: string;
  /**
   * Name of the destination cluster which can be used instead of server (url) field
   */
  name: string;
}

export interface OrphanedResource {
  group: string;
  kind: string;
  name: string;
}

export interface ApplicationSource {
  targetRevision: string;
  /**
   * RepoURL is repository URL which contains application project.
   */
  repoURL: string;

  /**
   * Path is a directory path within repository which
   */
  path?: string;

  chart?: string;

  helm?: ApplicationSourceHelm;

  kustomize?: ApplicationSourceKustomize;

  plugin?: ApplicationSourcePlugin;

  directory?: ApplicationSourceDirectory;
}
// ...
export interface SyncPolicy {
  automated?: Automated;
  syncOptions?: string[];
  retry?: RetryStrategy;
}

export interface Info {
  name: string;
  value: string;
}

export interface ApplicationSpec {
  project: string;
  source: ApplicationSource;
  sources: ApplicationSource[];
  destination: ApplicationDestination;
  syncPolicy?: SyncPolicy;
  ignoreDifferences?: ResourceIgnoreDifferences[];
  info?: Info[];
  revisionHistoryLimit?: number;
}

export interface ResourceIgnoreDifferences {
  group: string;
  kind: string;
  name: string;
  namespace: string;
  jsonPointers: string[];
  jqPathExpressions: string[];
}

/**
 * RevisionHistory contains information relevant to an application deployment
 */
export interface RevisionHistory {
  id: number;
  revision: string;
  source: ApplicationSource;
  revisions: string[];
  sources: ApplicationSource[];
  deployStartedAt: models.Time;
  deployedAt: models.Time;
}
// ...
export interface ResourceStatus {
  group: string;
  version: string;
  kind: string;
  namespace: string;
  name: string;
  status: SyncStatusCode;
  health: HealthStatus;
  createdAt?: models.Time;
  hook?: boolean;
  requiresPruning?: boolean;
  syncWave?: number;
}
// ...
export interface SyncStatus {
  comparedTo: ApplicationSource;
  status: SyncStatusCode;
  revision: string;
  revisions: string[];
}

export interface ApplicationCondition {
  type: string;
  message: string;
  lastTransitionTime: string;
}

export interface ApplicationSummary {
  externalURLs?: string[];
  images?: string[];
}

export interface ApplicationStatus {
  observedAt: models.Time;
  resources: ResourceStatus[];
  sync: SyncStatus;
  conditions?: ApplicationCondition[];
  history: RevisionHistory[];
  health: HealthStatus;
  operationState?: OperationState;
  summary?: ApplicationSummary;
}